pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "learnforge"
version = "0.1.0"
description = "Bootstrap milestone for a git-backed teaching publication system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "Jinja2>=3.1.4",
  "pydantic>=2.11.0,<3.0.0",
  "PyYAML>=6.0.2",
  "typer>=0.16.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=4.1.0",
  "pytest>=8.3.0",
  "ruff>=0.11.0",
]

[project.scripts]
teach = "app.cli:app"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "UP",
  "B",
]
